+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Properly render home screen actions buttons and profile animation.
+- Close login and logout bottom sheet once logged in or out.
 
 ## [2026.7.1] - 2026-07-10